// iterpro-frontend/libs/shared/data-access/permissions/src/interfaces/iterpro-routes.interface.ts
export type IterproRoute =
	| '/dashboards'
	| '/dashboards/standings'
	| '/manager'
	| '/manager/tactics'
	| '/manager/video-gallery'
	| '/manager/drills'
	| '/manager/planning'
	| '/manager/attendances'
	| '/performance'
	| '/performance/session-analysis'
	| '/performance/workload-analysis'
	| '/performance/readiness'
	| '/performance/assessments'
	| '/performance/test-analysis'
	| '/medical'
	| '/medical/maintenance'
	| '/medical/infirmary'
	| '/medical/medical-statistics'
	| '/medical/examination'
	| '/medical/medical-test-analysis'
	| '/players'
	| '/players/my-team'
	| '/players/compare-players'
	| '/players/scouting'
	| '/players/transfers'
	| '/administration'
	| '/administration/squads'
	| '/administration/finance-overview'
	| '/administration/bonus'
	| '/administration/cash-flow'
	| '/administration/cost-items'
	| '/repository'
	| '/import-data'
	| '/inbox'
	| '/notifications'
	| '/settings'
	| '/settings/user-preferences/users-permissions'
	| '/settings/club-preferences'
	| '/settings/club-preferences/general'
	| '/settings/club-preferences/seasons'
	| '/settings/club-preferences/scouting'
	| '/settings/club-preferences/finance'
	| '/settings/teams/seasons'
	| '/settings/teams/go-score';